Improving performance by using float.

+++ Test/DiKErnel.FunctionLibrary.Test/AsphaltWaveImpact/AsphaltWaveImpactFunctionsTest.cs
@@ -29,68 +29,68 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
         public void IncrementDamage_ValidInput_ExpectedValue()
         {
             // Setup
-            (double, double)[] widthFactors =
+            (float, float)[] widthFactors =
             {
-                (0.1, 0.0392),
-                (0.2, 0.0738),
-                (0.3, 0.1002),
-                (0.4, 0.1162),
-                (0.5, 0.1213),
-                (0.6, 0.1168),
-                (0.7, 0.1051),
-                (0.8, 0.089),
-                (0.9, 0.0712),
-                (1, 0.0541),
-                (1.1, 0.0391),
-                (1.2, 0.0269),
-                (1.3, 0.0216),
-                (1.4, 0.015),
-                (1.5, 0.0105)
+                (0.1f, 0.0392f),
+                (0.2f, 0.0738f),
+                (0.3f, 0.1002f),
+                (0.4f, 0.1162f),
+                (0.5f, 0.1213f),
+                (0.6f, 0.1168f),
+                (0.7f, 0.1051f),
+                (0.8f, 0.089f),
+                (0.9f, 0.0712f),
+                (1, 0.0541f),
+                (1.1f, 0.0391f),
+                (1.2f, 0.0269f),
+                (1.3f, 0.0216f),
+                (1.4f, 0.015f),
+                (1.5f, 0.0105f)
             };
 
-            (double, double)[] depthFactors =
+            (float, float)[] depthFactors =
             {
-                (-1, 0.0244),
-                (-0.875, 0.0544),
-                (-0.75, 0.0938),
-                (-0.625, 0.1407),
-                (-0.5, 0.1801),
-                (-0.375, 0.1632),
-                (-0.25, 0.1426),
-                (-0.125, 0.0994),
-                (0, 0.06),
-                (0.125, 0.0244),
-                (0.25, 0.0169)
+                (-1, 0.0244f),
+                (-0.875f, 0.0544f),
+                (-0.75f, 0.0938f),
+                (-0.625f, 0.1407f),
+                (-0.5f, 0.1801f),
+                (-0.375f, 0.1632f),
+                (-0.25f, 0.1426f),
+                (-0.125f, 0.0994f),
+                (0, 0.06f),
+                (0.125f, 0.0244f),
+                (0.25f, 0.0169f)
             };
 
-            (double, double)[] impactFactors =
+            (float, float)[] impactFactors =
             {
-                (2, 0.039),
-                (2.4, 0.1),
-                (2.8, 0.18),
-                (3.2, 0.235),
-                (3.6, 0.2),
-                (4, 0.13),
-                (4.4, 0.08),
-                (4.8, 0.02),
-                (5.2, 0.01),
-                (5.6, 0.005),
-                (6, 0.001)
+                (2, 0.039f),
+                (2.4f, 0.1f),
+                (2.8f, 0.18f),
+                (3.2f, 0.235f),
+                (3.6f, 0.2f),
+                (4, 0.13f),
+                (4.4f, 0.08f),
+                (4.8f, 0.02f),
+                (5.2f, 0.01f),
+                (5.6f, 0.005f),
+                (6, 0.001f)
             };
 
-            const double logFlexuralStrength = 0.193124598;
-            const double averageNumberOfWaves = 6492.937853;
-            const double maximumPeakStress = 0.0160884;
-            const double stiffnessRelation = 1.185626183;
-            const double computationalThickness = 0.16;
-            const double outerSlope = 0.232914161;
+            const float logFlexuralStrength = 0.193124598f;
+            const float averageNumberOfWaves = 6492.937853f;
+            const float maximumPeakStress = 0.0160884f;
+            const float stiffnessRelation = 1.185626183f;
+            const float computationalThickness = 0.16f;
+            const float outerSlope = 0.232914161f;
 
-            const double z = 2.988;
-            const double waterLevel = 3.8;
-            const double waveHeightHm0 = 1.6;
-            const double fatigueAlpha = 0.42;
-            const double fatigueBeta = 4.76;
-            const double impactNumberC = 1;
+            const float z = 2.988f;
+            const float waterLevel = 3.8f;
+            const float waveHeightHm0 = 1.6f;
+            const float fatigueAlpha = 0.42f;
+            const float fatigueBeta = 4.76f;
+            const float impactNumberC = 1;
 
             var input = new AsphaltWaveImpactInput(logFlexuralStrength, averageNumberOfWaves, maximumPeakStress,
                                                    stiffnessRelation, computationalThickness, outerSlope, widthFactors,
@@ -98,7 +98,7 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
                                                    fatigueAlpha, fatigueBeta, impactNumberC);
 
             // Call
-            double incrementDamage = AsphaltWaveImpactFunctions.IncrementDamage(input);
+            float incrementDamage = AsphaltWaveImpactFunctions.IncrementDamage(input);
 
             // Assert
             AssertHelper.AreEqual(7.94813500019044, incrementDamage);
@@ -108,10 +108,10 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
         public void LogFlexuralStrength_ValidInput_ExpectedValue()
         {
             // Setup
-            const double flexuralStrength = 1.56;
+            const float flexuralStrength = 1.56f;
 
             // Call
-            double logFlexuralStrength = AsphaltWaveImpactFunctions.LogFlexuralStrength(flexuralStrength);
+            float logFlexuralStrength = AsphaltWaveImpactFunctions.LogFlexuralStrength(flexuralStrength);
 
             // Assert
             AssertHelper.AreEqual(0.19312459835446, logFlexuralStrength);
@@ -121,12 +121,12 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
         public void MaximumPeakStress_ValidInput_ExpectedValue()
         {
             // Setup
-            const double waveHeightHm0 = 1.6;
-            const double gravitationalAcceleration = 9.81;
-            const double densityOfWater = 1025;
+            const float waveHeightHm0 = 1.6f;
+            const float gravitationalAcceleration = 9.81f;
+            const float densityOfWater = 1025;
 
             // Call
-            double maximumPeakStress = AsphaltWaveImpactFunctions.MaximumPeakStress(
+            float maximumPeakStress = AsphaltWaveImpactFunctions.MaximumPeakStress(
                 waveHeightHm0, gravitationalAcceleration, densityOfWater);
 
             // Assert
@@ -137,13 +137,13 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
         public void StiffnessRelation_ValidInput_ExpectedValue()
         {
             // Setup
-            const double computationalThickness = 0.16;
-            const double equivalentElasticModulus = 18214;
-            const double soilElasticity = 56;
-            const double stiffnessRelationNu = 0.35;
+            const float computationalThickness = 0.16f;
+            const float equivalentElasticModulus = 18214;
+            const float soilElasticity = 56;
+            const float stiffnessRelationNu = 0.35f;
 
             // Call
-            double stiffnessRelation = AsphaltWaveImpactFunctions.StiffnessRelation(
+            float stiffnessRelation = AsphaltWaveImpactFunctions.StiffnessRelation(
                 computationalThickness, equivalentElasticModulus, soilElasticity, stiffnessRelationNu);
 
             // Assert
@@ -154,13 +154,13 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
         public void ComputationalThickness_ValidInput_ExpectedValue()
         {
             // Setup
-            const double thicknessUpperLayer = 0.16;
-            const double thicknessSubLayer = 0.23;
-            const double elasticModulusUpperLayer = 18214;
-            const double elasticModulusSubLayer = 19214;
+            const float thicknessUpperLayer = 0.16f;
+            const float thicknessSubLayer = 0.23f;
+            const float elasticModulusUpperLayer = 18214;
+            const float elasticModulusSubLayer = 19214;
 
             // Call
-            double computationalThickness = AsphaltWaveImpactFunctions.ComputationalThickness(
+            float computationalThickness = AsphaltWaveImpactFunctions.ComputationalThickness(
                 thicknessUpperLayer, thicknessSubLayer, elasticModulusUpperLayer, elasticModulusSubLayer);
 
             // Assert
@@ -171,13 +171,13 @@ namespace DiKErnel.FunctionLibrary.Test.AsphaltWaveImpact
         public void OuterSlope_ValidInput_ExpectedValue()
         {
             // Setup
-            const double slopeLowerPosition = 2.4;
-            const double slopeLowerLevel = 1.75;
-            const double slopeUpperPosition = 10.2;
-            const double slopeUpperLevel = 2.95;
+            const float slopeLowerPosition = 2.4f;
+            const float slopeLowerLevel = 1.75f;
+            const float slopeUpperPosition = 10.2f;
+            const float slopeUpperLevel = 2.95f;
 
             // Call
-            double outerSlope = AsphaltWaveImpactFunctions.OuterSlope(
+            float outerSlope = AsphaltWaveImpactFunctions.OuterSlope(
                 slopeLowerPosition, slopeLowerLevel, slopeUpperPosition, slopeUpperLevel);
 
             // Assert
